mandacaru_bot/requirements.txt
# ===============================================
# DEPENDÊNCIAS DO BOT TELEGRAM MANDACARU
# ===============================================

# Bot Telegram
aiogram==3.4.1

# HTTP Client para API
httpx==0.26.0

# Configuração
python-dotenv==1.0.0

# Monitoramento de sistema
psutil==5.9.8

# Opcional: Para melhor performance
aiofiles==23.2.1